feat(provider): add Prime Agent ACP integration
