+++ hermes_cli/model_switch.py
@@ -2423,6 +2423,16 @@ def list_authenticated_providers(
                 has_creds = has_vertex_credentials()
             except Exception as exc:
                 logger.debug("Vertex credential check failed: %s", exc)
+        elif overlay.auth_type == "external_process":
+            # Local subprocess-backed providers (for example Prime Agent)
+            # authenticate by the executable being available, not by an env key.
+            try:
+                from hermes_cli.auth import get_external_process_provider_status
+                has_creds = bool(
+                    get_external_process_provider_status(hermes_slug).get("configured")
+                )
+            except Exception as exc:
+                logger.debug("External-process provider check failed for %s: %s", hermes_slug, exc)
         elif overlay.extra_env_vars:
             has_creds = any(os.environ.get(ev) for ev in overlay.extra_env_vars)
         # Also check api_key_env_vars from PROVIDER_REGISTRY for api_key auth_type
@@ -2493,7 +2503,7 @@ def list_authenticated_providers(
         if not has_creds:
             continue
 
-        if hermes_slug in {"openai-codex", "copilot", "copilot-acp"}:
+        if hermes_slug in {"openai-codex", "copilot", "copilot-acp", "prime-agent"}:
             # Use live OAuth-backed discovery so the gateway /model picker
             # matches what the user's authenticated Codex/Copilot backend
             # actually serves — including ChatGPT-Pro-only Codex slugs
