+++ hermes_cli/runtime_provider.py
@@ -2025,12 +2025,13 @@ def resolve_runtime_provider(
                 "requested_provider": requested_provider,
             }
 
-    if provider == "copilot-acp":
+    if provider in {"copilot-acp", "prime-agent"}:
         creds = resolve_external_process_provider_credentials(provider)
         return {
-            "provider": "copilot-acp",
+            "provider": provider,
             "api_mode": "chat_completions",
-            "base_url": creds.get("base_url", "").rstrip("/"),
+            "base_url": creds.get("base_url", "").rstrip("/")
+            or ("acp://prime-agent" if provider == "prime-agent" else "acp://copilot"),
             "api_key": creds.get("api_key", ""),
             "command": creds.get("command", ""),
             "args": list(creds.get("args") or []),
