+++ tests/run_agent/test_streaming.py
@@ -1292,66 +1292,28 @@ def _make_acp_agent(provider="copilot-acp", base_url="acp://copilot"):
     return agent
 
 
-class TestCopilotACPStreamingDecision:
-    """Verify that copilot-acp routes to the non-streaming path.
+class TestACPStreamingDecision:
+    """Local ACP subprocess providers expose iterable live streams."""
 
-    CopilotACPClient communicates via subprocess stdio and returns a plain
-    SimpleNamespace — not an iterable stream.  The streaming decision logic
-    must detect ACP runtimes and route to _interruptible_api_call instead.
-    """
-
-    @patch("run_agent.get_tool_definitions", return_value=[])
-    @patch("run_agent.check_toolset_requirements", return_value={})
-    @patch("agent.copilot_acp_client.CopilotACPClient")
-    def test_provider_name_triggers_non_streaming(
-        self, mock_acp_cls, _mock_check, _mock_tools
-    ):
-        """provider='copilot-acp' → non-streaming path."""
-        mock_acp_cls.return_value = MagicMock()
-        agent = _make_acp_agent(provider="copilot-acp", base_url="acp://copilot")
-
-        with (
-            patch.object(agent, "_interruptible_api_call",
-                         return_value=_valid_acp_response()) as mock_non_stream,
-            patch.object(agent, "_interruptible_streaming_api_call") as mock_stream,
-        ):
-            # Verify the decision logic correctly disables streaming
-            _use_streaming = True
-            if getattr(agent, "_disable_streaming", False):
-                _use_streaming = False
-            elif (
-                agent.provider == "copilot-acp"
-                or str(agent.base_url or "").lower().startswith("acp://copilot")
-                or str(agent.base_url or "").lower().startswith("acp+tcp://")
-            ):
-                _use_streaming = False
-
-            assert _use_streaming is False
-            # Call the non-streaming path as the loop would
-            response = mock_non_stream({})
-            mock_stream.assert_not_called()
-
-    @patch("run_agent.get_tool_definitions", return_value=[])
-    @patch("run_agent.check_toolset_requirements", return_value={})
-    @patch("agent.copilot_acp_client.CopilotACPClient")
-    def test_acp_base_url_triggers_non_streaming(
-        self, mock_acp_cls, _mock_check, _mock_tools
-    ):
-        """base_url='acp://copilot' → non-streaming even without provider name."""
-        mock_acp_cls.return_value = MagicMock()
-        agent = _make_acp_agent(provider="custom", base_url="acp://copilot")
-        agent.provider = "custom"
-
+    def test_local_acp_provider_allows_streaming(self):
+        agent = _make_acp_agent(provider="prime-agent", base_url="acp://prime-agent")
         _use_streaming = True
-        if (
-            agent.provider == "copilot-acp"
-            or str(agent.base_url or "").lower().startswith("acp://copilot")
-            or str(agent.base_url or "").lower().startswith("acp+tcp://")
-        ):
+        if getattr(agent, "_disable_streaming", False):
+            _use_streaming = False
+        elif str(agent.base_url or "").lower().startswith("acp+tcp://"):
+            _use_streaming = False
+
+        assert _use_streaming is True
+
+    def test_remote_acp_tcp_stays_non_streaming(self):
+        agent = _make_acp_agent(provider="custom", base_url="acp+tcp://127.0.0.1:9999")
+        _use_streaming = True
+        if getattr(agent, "_disable_streaming", False):
+            _use_streaming = False
+        elif str(agent.base_url or "").lower().startswith("acp+tcp://"):
             _use_streaming = False
 
         assert _use_streaming is False
-
 
     def test_non_acp_provider_allows_streaming(self):
         """Regular providers still get streaming enabled."""
